tests/wpt/web-platform-tests/referrer-policy/css-integration/image/internal-import-stylesheet.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - image from imported stylesheet (internal)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/referrer-policy/generic/common.js"></script>
    <meta name="referrer" content="origin">
  </head>
  <body>
    <p>Check that resources from stylesheets (imported from internal
    stylesheets) are loaded with the referrer and referrer policy from the
    document.</p>

    <div class="styled"></div>

    <script>
      async_test(function(css_test) {
        var id = token();
        var url_prefix =  location.protocol + "//www1." + location.hostname + ":" + location.port + "/referrer-policy/generic/subresource/";
        var css_url = url_prefix + "stylesheet.py?id=" + id;
        var img_url = url_prefix + "image.py?report-headers&id=" + id;

        var style = document.createElement("style");
        style.type = 'text/css';
        style.appendChild(document.createTextNode("@import url('" + css_url + "');"));
        document.head.appendChild(style);
        css_test.step_timeout(function() {
            queryXhr(img_url, function(message) {
                assert_own_property(message, "headers");
                assert_own_property(message, "referrer");
                assert_equals(message.referrer, css_url);
                css_test.done();
            }, null, null, css_test);
        }, 1000);
      }, "Image from imported stylesheet (internal).");
    </script>

    <div id="log"></div>
  </body>
</html>
